Refactor Array instruction classes
